Made all hotkeys configurable
Notable exceptions are ctrl+up/down and pgup/pgdown for navigation in
index viewer, tab as autocompletion, and obvious ones as ctrl-c/v etc in
the various text fields.
Also the option to have multiple hotkey alternatives for one action is
now gone. (Making hotkeys configurable would have been a mess with that
in place)
Closes #15 (FINALLY it only took me little over two years sheesh)